+++ src/test/java/org/delia/bdd/AllBDDTests.java
@@ -328,7 +328,7 @@ public class AllBDDTests extends BDDBase {
 //		runR1500File("t0-queryfn-orderby-2span.txt", 1);
 //		runR1500File("t0-queryfn-orderby.txt", 4);
 //		runR560File("t0-self-NN.txt", 4);
-		runR1600File("t0-fetch.txt", 3);
+		runR1300File("t0-let-query.txt", 7);
 
 	}
